Move rotation Lambda POC to dedicated module
Resolves: #19
