+++ fernet-aws-secrets-manager-rotator/src/main/java/com/macasaet/fernet/aws/secretsmanager/rotation/SimpleFernetKeyRotator.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package com.macasaet.fernet.example.rotation.aws;
+package com.macasaet.fernet.aws.secretsmanager.rotation;
 
 import static java.util.Collections.singletonList;
